+++ src/manifest.js
@@ -47,11 +47,19 @@ Manifest.prototype = {
     ))
   },
 
-  createMiddlewares() {
+  /**
+   * @param {Object} args
+   *   @param {String} args.resourceName
+   *   @param {String} args.resourceMethod
+   *   @param {Boolean} args.mockRequest
+   *
+   * @return {Array<Object>}
+   */
+  createMiddlewares(args = {}) {
     const createInstance = (middlewareFactory) => assign({
       request: (request) => request,
       response: (next) => next()
-    }, middlewareFactory())
+    }, middlewareFactory(args))
 
     return this.middlewares
       .map((middleware) => createInstance(middleware))